metiq: clean up of input parameters and code location
More cleanup is needed and the parameters simplified.
