+++ src/metiq.py
@@ -309,7 +309,6 @@ def get_options(argv):
     parser.add_argument(
         "-i",
         "--input",
-        nargs="+",
         type=str,
         dest="infile",
         default=default_values["infile"],
@@ -517,36 +516,20 @@ def main(argv):
     elif options.func == "analyze":
         media_analyze.media_analyze(
             analysis_type=options.analysis_type,
-            width=options.width,
-            height=options.height,
-            num_frames=options.num_frames,
-            pixel_format=options.pixel_format,
-            luma_threshold=options.luma_threshold,
             pre_samples=options.pre_samples,
             samplerate=options.samplerate,
             beep_freq=options.beep_freq,
             beep_duration_samples=options.beep_duration_samples,
             beep_period_sec=options.beep_period_sec,
             scale=options.scale,
-            infile=options.infile,
             input_video=options.input_video,
             input_audio=options.input_audio,
             outfile=options.outfile,
-            vft_id=options.vft_id,
-            cache_video=options.cache_video,
-            cache_audio=options.cache_audio,
-            cache_both=options.cache_both,
-            min_separation_msec=options.min_separation_msec,
-            min_match_threshold=options.min_match_threshold,
             audio_sample=options.audio_sample,
-            lock_layout=options.lock_layout,
-            tag_manual=options.tag_manual,
             force_fps=options.force_fps,
-            threaded=options.threaded,
             audio_offset=options.audio_offset,
             z_filter=options.z_filter,
             windowed_stats_sec=options.windowed_stats_sec,
-            no_hw_decode=options.no_hw_decode,
             debug=options.debug,
         )
